[REF] html_editor: split color plugin in two
This commit splits the current color plugin into two plugins:
- ColorPlugin: contains the code to get and apply colors to elements or
selections. This is a technical plugin that other plugins can depend on
to handle color-related formatting.
- ColorUIPlugin: provides a user interface for color formatting by
adding the color selector to the toolbar. This plugin holds the state
needed for the color selector and its previewable operations.

Other than improving modularity, this split allows us to have an editor
without the color selector, for the cases where we don't want any color
formatting by the user.

task-5043928

X-original-commit: 06e67bc886865d6ee75ef1f4c0ca58b45ebbb624
