+++ addons/html_editor/static/src/main/media/icon_plugin.js
@@ -8,7 +8,7 @@ import { ICON_SELECTOR, isElement } from "@html_editor/utils/dom_info";
 
 export class IconPlugin extends Plugin {
     static id = "icon";
-    static dependencies = ["history", "selection", "color", "dialog"];
+    static dependencies = ["history", "selection", "colorUi", "dialog"];
     resources = {
         user_commands: [
             {
@@ -80,7 +80,7 @@ export class IconPlugin extends Plugin {
                 groupId: "icon_color",
                 description: _t("Select Font Color"),
                 Component: ColorSelector,
-                props: this.dependencies.color.getPropsForColorSelector("foreground"),
+                props: this.dependencies.colorUi.getPropsForColorSelector("foreground"),
                 isAvailable: isHtmlContentSupported,
             },
             {
@@ -88,7 +88,7 @@ export class IconPlugin extends Plugin {
                 groupId: "icon_color",
                 description: _t("Select Background Color"),
                 Component: ColorSelector,
-                props: this.dependencies.color.getPropsForColorSelector("background"),
+                props: this.dependencies.colorUi.getPropsForColorSelector("background"),
                 isAvailable: isHtmlContentSupported,
             },
             {
