+++ addons/html_editor/static/src/plugin_sets.js
@@ -23,6 +23,7 @@ import { ChatGPTTranslatePlugin } from "./main/chatgpt/chatgpt_translate_plugin"
 import { ColumnPlugin } from "./main/column_plugin";
 import { EmojiPlugin } from "./main/emoji_plugin";
 import { ColorPlugin } from "./main/font/color_plugin";
+import { ColorUIPlugin } from "./main/font/color_ui_plugin";
 import { FeffPlugin } from "./main/feff_plugin";
 import { FontPlugin } from "./main/font/font_plugin";
 import { FontFamilyPlugin } from "./main/font/font_family_plugin";
@@ -98,6 +99,7 @@ import { ContentEditablePlugin } from "./core/content_editable_plugin";
  *
  * Main
  * @property { import("./main/font/color_plugin").ColorShared } color
+ * @property { import("./main/font/color_ui_plugin").ColorUIShared } colorUi
  * @property { import("./main/link/link_plugin").LinkShared } link
  * @property { import ("./main/link/link_selection_plugin").LinkSelectionShared } linkSelection
  * @property { import ("./main/media/media_plugin").MediaShared } media
@@ -147,6 +149,7 @@ export const MAIN_PLUGINS = [
     BannerPlugin,
     ChatGPTTranslatePlugin,
     ColorPlugin,
+    ColorUIPlugin,
     SeparatorPlugin,
     ColumnPlugin,
     EmojiPlugin,
